Fix memory error, bring back CI tests.
I commented out/xdescribed some failing tests until those groups can bring them
back.

I believe the memory leak was coming from holding references to Globals, which
hold references back to WASM modules, which hold references to memories, and
there can only be so many WASM memories allocated at a time (NOT because you
run out of memory in process, but because you run out of virtual address
space).

Here's a short program that demos this:

mems = [];
function makeMemories(n, size) {
    for(var i = 0; i < n; i += 1) {
        const m = new WebAssembly.Memory({ initial: size });
        const view = new Int32Array(m.buffer);
        view[0] = i;
        mems.push(m);
    }
}

makeMemories(1000, 1000);
console.log("Memories made");

⤇ time node tests/makememories.js
/Users/joe/src/chocopy-wasm-compiler-B/tests/makememories.js:4
                const m = new WebAssembly.Memory({ initial: size });
                          ^

RangeError: WebAssembly.Memory(): could not allocate memory
    at makeMemories (/Users/joe/src/chocopy-wasm-compiler-B/tests/makememories.js:4:13)
    at Object.<anonymous> (/Users/joe/src/chocopy-wasm-compiler-B/tests/makememories.js:11:1)

Crucially, if we _don't_ store the memories in the `mems` array, this finishes
successfully. This means that garbage collection works fine with these
memories.

Now, which group's code would retain references to globals?

https://knowyourmeme.com/memes/were-all-trying-to-find-the-guy-who-did-this

That is, there was a place where the starter code stores the globals in an
`env` field on importObject. @zhs1397 even helpfully pointed out that it was
unused:

https://github.com/ucsd-cse231-s22/chocopy-wasm-compiler-B/commit/b4257cdc95cedae9964e947a178867598fa38242#diff-4920e1737cb25b31d048a90d1c0d03e5a08d0af92749356c528c0aaf362834ceR61

This might have been inevitable (?) on compiler A as well, but it's not
happening at all there. I'm not sure if there's a particular